Tree: invoke _ev_routing() on relaunch and add tests
EventHandler._ev_routing() has been defined since CS 1.6 or so, but
was never called. Testing the issue reported as #566 is an opportunity
to implement it for redistribute event. This allows us to test the
redistribution of commands on gateway failure, indeed we can now create
a test with a first gateway failing followed by a functional gateway
after redistribution.